recent_topics: Display compose box & enable compose hotkeys.
We move compose from being a part of message feed to
being a part of  middle column which is a common parent of recent
topics and message feed. This allows us to use a common compose
box for both the views. Fortunately, compose actions were
independent of this change so there weren't any evident
side effects.

We hide reply button in recent topics view for now. We may
bring it back with the ability to reply to selected topic / stream
after we are confident with bringing compose to recent topics.

We also enable compose hotkeys except for reply.

Fixes #17543