recent_topics: Scroll the focused topic in center.
Topics at top and bottom can sometimes be not visible due
to browser expecting them to be visible but they are not
due to some element visible above them with a higher z-index or
due to similar reasons.

This ensures that focused topic row is always visible by focusing
it in the center much like how messages view works.

We no longer want new rows to render after user is almost at
the last topic row, so load more topic much before that.